# wafconsole/api/wafTop/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /app/wafTop/v1/buildinRule/{id}:
        get:
            tags:
                - BuildRule
            operationId: BuildRule_GetBuildRule
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.wafTop.v1.GetBuildRuleReply'
    /app/wafTop/v1/buildinRules:
        get:
            tags:
                - BuildRule
            operationId: BuildRule_ListBuildRule
            parameters:
                - name: name
                  in: query
                  schema:
                    type: string
                - name: page_now
                  in: query
                  schema:
                    type: string
                - name: page_size
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.wafTop.v1.ListBuildRuleReply'
...
    /app/wafTop/v1/userRule:
        put:
            tags:
                - UserRule
            operationId: UserRule_UpdateUserRule
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.wafTop.v1.UpdateUserRuleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.wafTop.v1.UpdateUserRuleReply'
        post:
            tags:
                - UserRule
            operationId: UserRule_CreateUserRule
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.wafTop.v1.CreateUserRuleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.wafTop.v1.CreateUserRuleReply'
        delete:
            tags:
                - UserRule
            operationId: UserRule_DeleteUserRule
            parameters:
                - name: ids
                  in: query
                  schema:
                    type: array
                    items:
                        type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.wafTop.v1.DeleteUserRuleReply'
...
    /app/waf_top/v1/notice:
        post:
            tags:
                - Notice
            operationId: Notice_CreateNotice
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.wafTop.v1.CreateNoticeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.wafTop.v1.CreateNoticeReply'
        delete:
            tags:
                - Notice
            operationId: Notice_DeleteNotice
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.wafTop.v1.DeleteNoticeReply'
    /app/waf_top/v1/notice/state:
        put:
            tags:
                - Notice
            operationId: Notice_UpdateNotice
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.wafTop.v1.UpdateNoticeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.wafTop.v1.UpdateNoticeReply'
    /app/waf_top/v1/notice/{id}:
        get:
            tags:
                - Notice
            operationId: Notice_GetNotice
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.wafTop.v1.GetNoticeReply'
...
components:
    schemas:
...
        api.wafTop.v1.BuildinRule:
            type: object
            properties:
                id:
                    type: string
                risk_level:
                    type: string
                name:
                    type: string
                description:
                    type: string
                group_id:
                    type: string
                created_at:
                    type: string
                updated_at:
                    type: string
...
        api.wafTop.v1.CreateNoticeReply:
            type: object
            properties: {}
        api.wafTop.v1.CreateNoticeRequest:
            type: object
            properties: {}
            description: '##### notice'
...
        api.wafTop.v1.CreateUserRuleReply:
            type: object
            properties: {}
        api.wafTop.v1.CreateUserRuleRequest:
            type: object
            properties:
                name:
                    type: string
                description:
                    type: string
                risk_level:
                    type: string
                group_id:
                    type: string
                seclang_mod:
                    $ref: '#/components/schemas/api.wafTop.v1.SeclangMod'
            description: '########## 用户自定义规则 #########'
...
        api.wafTop.v1.DeleteNoticeReply:
            type: object
            properties: {}
...
        api.wafTop.v1.DeleteUserRuleReply:
            type: object
            properties: {}
...
        api.wafTop.v1.GetBuildRuleReply:
            type: object
            properties:
                name:
                    type: string
                description:
                    type: string
                risk_level:
                    type: string
                group_id:
                    type: string
        api.wafTop.v1.GetNoticeReply:
            type: object
            properties: {}
...
        api.wafTop.v1.ListBuildRuleReply:
            type: object
            properties:
                total:
                    type: string
                buildin_rules:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.wafTop.v1.BuildinRule'
...
        api.wafTop.v1.SeclangMod:
            type: object
            properties:
                match_goal:
                    type: string
                match_action:
                    type: string
                match_content:
                    type: string
...
        api.wafTop.v1.UpdateNoticeReply:
            type: object
            properties: {}
        api.wafTop.v1.UpdateNoticeRequest:
            type: object
            properties: {}
...
        api.wafTop.v1.UpdateUserRuleReply:
            type: object
            properties: {}
        api.wafTop.v1.UpdateUserRuleRequest:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                description:
                    type: string
                risk_level:
                    type: string
                group_id:
                    type: string
                seclang_mod:
                    $ref: '#/components/schemas/api.wafTop.v1.SeclangMod'
...